Add IDs and classes to semantic elements
Add IDs and classes to semantic elements so as to select them later for
styling. This is to allow for future flexibility, as targeting semantic
elements directly is not the best practice. By doing it this way, it
allows for semantic elements to be nested (e.g. you can use more than
one header or main elements throughout the webpage structure).