Set font sizes and colors to correct values
Set the font size and colors to the correct values. Also separate
style from layout in the stylesheet to make it clearer. Some vertical
margin was also added to the figcaption element in the quote section,
to improve visual appearance. Future work is needed to make some of
the CSS rules more general, so that they can be reused for future pages
this website may have.
work is needed to make the CSS rules more general, but